migrations fixed

+++ database/migrations/2026_02_04_164604_add_question_column_in_automated_messages_table.php
@@ -11,10 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('automated_messages', function (Blueprint $table) {
-            $table->string('question_for')->default(CUSTOMER)->after('id');
-            $table->string('question')->after('question_for');
-        });
+        if (! Schema::hasColumn('automated_messages', 'question_for')) {
+            Schema::table('automated_messages', function (Blueprint $table) {
+                $table->string('question_for')->default('customer')->after('id');
+            });
+        }
+
+        if (! Schema::hasColumn('automated_messages', 'question')) {
+            Schema::table('automated_messages', function (Blueprint $table) {
+                $table->string('question')->after('question_for');
+            });
+        }
     }
 
     /**
@@ -22,9 +29,16 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('automated_messages', function (Blueprint $table) {
-            $table->dropColumn('question_for');
-            $table->dropColumn('question');
-        });
+        if (Schema::hasColumn('automated_messages', 'question')) {
+            Schema::table('automated_messages', function (Blueprint $table) {
+                $table->dropColumn('question');
+            });
+        }
+
+        if (Schema::hasColumn('automated_messages', 'question_for')) {
+            Schema::table('automated_messages', function (Blueprint $table) {
+                $table->dropColumn('question_for');
+            });
+        }
     }
 };
